task3/src/requirements.txt
flake8==3.9.0
pre-commit==2.15.0
wemake-python-styleguide==0.15.3
flake8-bandit==2.1.2 
bandit==1.7.2
fastapi==0.75.2
orjson==3.6.8
pydantic[dotenv]==1.9.1
uvicorn==0.17.6
gunicorn==20.1.0
fastapi-jwt-auth==0.5.0
sqlalchemy==1.4.36
psycopg2-binary==2.9.3